refactor: 지원현황/면접조회 Board > 테이블 컴포넌트 분리
